requirements.txt
asgiref<=3.5.2
backports.zoneinfo<=0.2.1
boto3<=1.40.38
botocore<=1.40.38
Django<=5.2.6
django-cors-headers<=4.9.0
django-rest-framework<=0.1.0
django-storages<=1.14.6
djangorestframework<=3.16.1
djangorestframework-jwt<=1.11.0
djangorestframework-simplejwt<=5.5.1
dotenv==0.9.9
gunicorn<=20.1.0
jmespath<=1.0.1
Pillow<=9.3.0
psycopg2-binary<=2.9.5
PyJWT<=1.7.1
pyodbc<=4.0.35
python-dateutil<=2.8.2
pytz<=2022.7
s3transfer<=0.14.0
six<=1.16.0
sqlparse<=0.4.3
tzdata<=2022.7
urllib3<=1.26.13
whitenoise<=6.11.0
wrapt<=1.12.1
mysqlclient<=2.2.7
